# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/users/accounts": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает список аккаунтов Valorant, принадлежащих авторизованному пользователю.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Получение списка аккаунтов пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Максимальное количество записей",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение для пагинации",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список аккаунтов успешно получен",
                        "schema": {
                            "$ref": "#/definitions/users.getAccountsResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/users.getAccountsResponse"
                        }
                    }
                }
            }
        },
        "/v1/users/login": {
            "post": {
                "description": "Выполняет вход пользователя по логину и паролю, устанавливает cookies авторизации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Аутентификация пользователя",
                "parameters": [
                    {
                        "description": "Данные для входа пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.authUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь успешно авторизован",
                        "schema": {
                            "$ref": "#/definitions/users.authUserResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные в запросе",
                        "schema": {
                            "$ref": "#/definitions/users.authUserResponse"
                        }
                    },
                    "401": {
                        "description": "Неверный логин или пароль",
                        "schema": {
                            "$ref": "#/definitions/users.authUserResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/users.authUserResponse"
                        }
                    }
                }
            }
        },
        "/v1/users/logout": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Удаляет refresh токен из Redis и очищает cookies авторизации",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Выход из системы",
                "responses": {
                    "200": {
                        "description": "Выход выполнен успешно",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/v1/users/me": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает данные авторизованного пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Информация о текущем пользователе",
                "responses": {
                    "200": {
                        "description": "Данные пользователя",
                        "schema": {
                            "$ref": "#/definitions/users.meResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/users.meResponse"
                        }
                    }
                }
            }
        },
        "/v1/users/refresh": {
            "post": {
                "description": "Обновляет access токен используя refresh токен из cookie. Ротирует refresh токен.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Обновление access токена",
                "responses": {
                    "200": {
                        "description": "Токен успешно обновлен",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Refresh токен отсутствует или недействителен",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/v1/users/register": {
            "post": {
                "description": "Создает нового пользователя с указанным логином и паролем, устанавливает cookies авторизации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Регистрация нового пользователя",
                "parameters": [
                    {
                        "description": "Данные для регистрации пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.authUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Пользователь успешно зарегистрирован",
                        "schema": {
                            "$ref": "#/definitions/users.authUserResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные в запросе",
                        "schema": {
                            "$ref": "#/definitions/users.authUserResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера при создании пользователя",
                        "schema": {
                            "$ref": "#/definitions/users.authUserResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {},
                "message": {
                    "type": "string"
                },
                "meta": {}
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorResponse"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "users.accountDTO": {
            "type": "object",
            "properties": {
                "game_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "puuid": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "rr": {
                    "type": "integer"
                },
                "shard": {
                    "type": "string"
                },
                "tag_line": {
                    "type": "string"
                },
                "tier": {
                    "type": "integer"
                }
            }
        },
        "users.authUserRequest": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "users.authUserResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorResponse"
                },
                "login": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "users.getAccountsResponse": {
            "type": "object",
            "properties": {
                "accounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/users.accountDTO"
                    }
                },
                "error": {
                    "$ref": "#/definitions/response.ErrorResponse"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "users.meResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorResponse"
                },
                "id": {
                    "type": "integer"
                },
                "login": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}